Delay sending messages by some arbitrary timespan (20 milliseconds)
This is really gross and is hopefully just a temporary solution.
The problem is that `IObservable<JObject> Send<T>(string, T)` sends
a message to the server then immediately returns its observable,
which is then subscribed to. But the message may be sent and its
response received before the return value is subscribed to (at
least this can happen locally--it may never happen in production).